pyproject.toml
[project]
name = "assesment2"
version = "0.1.0"
description = "FastAPI + MongoDB Example"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pymongo",
    "python-dotenv",
    "pydantic",
    "pytest",
    "httpx"
]

[tool.uv]
# keep venv inside project for CI/CD simplicity
venv = ".venv"
